CapD/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
openai
qdrant-client
python-dotenv